% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{apply.formula.HLM}
\alias{apply.formula.HLM}
\title{Apply Hierarchical Linear Model with specific formula}
\usage{
apply.formula.HLM(
  r,
  X,
  Y,
  MARGIN = 1,
  formula = "y ~ (1 | samples) + x",
  ttest.flag = F
)
}
\arguments{
\item{r}{list object of sequencing data}

\item{X}{fixed effects}

\item{Y}{outcome}

\item{formula}{formula for the hierarchical model}

\item{ttest.flag}{(default = F) is flag for computing ttest?}
}
\value{
pvalues of hlm
}
\description{
Wrapper to compute differentially expressed genes with
a hierarchical linear model
}
