% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{cap_object}
\alias{cap_object}
\title{Cap Data}
\usage{
cap_object(X, q)
}
\arguments{
\item{X}{a list-list object, matrix, dataframe, or list}

\item{q}{quantile to cap at (e.g. 0.05 is capping at the bottom 5th and top 95th quantiles)}
}
\value{
a capped version of your input object.
}
\description{
Cap the values of data (matrix, data.frame, or list)
}
