% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{center.matrix}
\alias{center.matrix}
\title{Matrix centering}
\usage{
center.matrix(m, dim = 1, sd.flag = F)
}
\arguments{
\item{m}{matrix object}

\item{dim}{which dimension, rows = 1, columns = 2}

\item{sd.flag}{boolean for standardizing data to sd = 1 (default = F)}
}
\value{
a matrix of zscores
}
\description{
Code to center your matrix
}
