% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{cluster}
\alias{cluster}
\title{Wrapper around sNN clustering in Seurat}
\usage{
cluster(so, resolution = 0.4)
}
\arguments{
\item{so}{a SeuratObject of your data}

\item{resolution}{clustering resolution, default = 0.4}
}
\value{
a list of all parameters calculated
}
\description{
functionality for returning clusters
}
