% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{discretize.prvt}
\alias{discretize.prvt}
\title{Discretize data into Quantiles}
\usage{
discretize.prvt(v, n.cat, q1)
}
\arguments{
\item{v}{vector of values}

\item{n.cat}{number of bins}

\item{q1}{quantiles (not used anymore?)}
}
\value{
discretized data
}
\description{
discretizes a vector of values
}
