% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{embed}
\alias{embed}
\title{Wrapper around UMAP and/or TSNE}
\usage{
embed(so, tsne.flag = F, umap.flag = F, n_pcs = 30)
}
\arguments{
\item{so}{a SeuratObject of your data}

\item{tsne.flag}{boolean, run tSNE?}

\item{umap.flag}{boolean, run umap?}

\item{n_pcs}{number of PCs to run umap and tsne on}
}
\value{
a list of all parameters calculated
}
\description{
functionality for returning UMAP an TSNE results
}
