% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{formula.HLM}
\alias{formula.HLM}
\title{Execute lmer using formula}
\usage{
\method{formula}{HLM}(
  y,
  x,
  r0,
  formula = "y ~ (1 | samples) + x",
  val = ifelse(is.numeric(x), "", "TRUE"),
  return.all = F
)
}
\arguments{
\item{y}{list object of sequencing data}

\item{x}{fixed effects}

\item{r0}{outcome}

\item{formula}{formula for the hierarchical model}

\item{ttest.flag}{(default = F) is flag for computing ttest?}
}
\value{
pvalues of hlm
}
\description{
Wrapper to compute with hierarchical
linear model
}
