% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_assignments.R
\name{get.binary.scores}
\alias{get.binary.scores}
\title{Get Binary Scores}
\usage{
get.binary.scores(x1, EM.flag)
}
\arguments{
\item{x1}{list of the markers score for a particular cell-type}
}
\value{
a list of binary scores
}
\description{
function for getting binary scores (tends to be more relevant)
for the EM algorithm
}
