% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_assignments.R
\name{get.mat}
\alias{get.mat}
\title{Get Matrix}
\usage{
get.mat(m.rows, m.cols, data = NA)
}
\arguments{
\item{m.rows}{matrix rows}

\item{m.cols}{matrix columns}

\item{data}{your data}
}
\value{
your data in matrix format
}
\description{
function for casting data into a matrix.
}
