% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{get.OE}
\alias{get.OE}
\title{Overall Expression Wrapper}
\usage{
get.OE(r, sig)
}
\arguments{
\item{r}{data list}

\item{sig}{(default = 50)}
}
\value{
r, your data list
}
\description{
wrapper around code that computes overall expression
signature for each cell-type
}
