% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{get.OE1}
\alias{get.OE1}
\title{Overall Expression Subroutine}
\usage{
get.OE1(r, sig)
}
\arguments{
\item{r}{data list}

\item{sig}{(default = 50)}
}
\value{
r, your data list
}
\description{
code that computes overall expression
signature for each cell-type
}
