% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{get.p.zscores}
\alias{get.p.zscores}
\title{Convert p-values to z-scores}
\usage{
get.p.zscores(p)
}
\arguments{
\item{p}{a table of p values where p\link{,1} is the greater than hypothesis, and the p\link{,2} is the less than hypothesis}
}
\value{
\emph{z}-scores
}
\description{
samples \emph{size} number of cells from each
labelled population
}
