% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{get.top.cor}
\alias{get.top.cor}
\title{Fetch genes with top correlation}
\usage{
get.top.cor(m, q = 100, min.ci = 0, idx = NULL, add.prefix = "")
}
\arguments{
\item{m}{matrix}

\item{q}{(default = 100)}

\item{min.ci}{(default = 0)}

\item{idx}{(default = NULL)}

\item{add.prefix}{= (default = "") string to prepend to names}
}
\description{
Wrapper around a rank fetch
}
