% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{get.top.elements}
\alias{get.top.elements}
\title{Fetch top elements of a matrix}
\usage{
get.top.elements(m, q = 100, min.ci = NULL, main = "")
}
\arguments{
\item{m}{matrix}

\item{q}{(default = 100)}

\item{min.ci}{(default = 0)}

\item{idx}{(default = NULL)}

\item{add.prefix}{= (default = "") string to prepend to names}
}
\description{
Wrapper around a rank fetch
}
