% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{intersect.list1}
\alias{intersect.list1}
\title{Intersect Lists of Lists}
\usage{
intersect.list1(l, g, n1 = 0, HG.universe = NULL, prf = "")
}
\arguments{
\item{l}{first list of lists}

\item{g}{second lists of lists}

\item{n1}{cutoff for length of list}

\item{HG.universe}{genes for computing GO enrichment}

\item{prf}{some sort of string parsing}
}
\value{
\emph{z}-scores
}
\description{
functionality for finding the intersection
of lists of lists.
}
