% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{recon_loss}
\alias{recon_loss}
\title{Reconstruction Loss}
\usage{
recon_loss(X, X_r)
}
\arguments{
\item{X}{original matrix \emph{n} x \emph{m} matrix}

\item{X_r}{low-dimensional reconstructed \emph{n} x \emph{m} matrix}
}
\value{
a list of reconstruction loss, labeled with cellnames
}
\description{
functionality for returning reconstruction loss
for any dimension reduction algorithm. Please make sure
that the the order of your rows and columns are the same!
}
