% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageProcessing.R
\name{rgb2hex}
\alias{rgb2hex}
\title{RGB to Hexcode}
\usage{
rgb2hex(x)
}
\arguments{
\item{x}{length three string corresponding to RGB color code}
}
\value{
a hexcode (string)
}
\description{
Simple function that converts a length 3 list corresponding
to RGB code to a color hexcode
}
