% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_wrappers.R
\name{run_pca}
\alias{run_pca}
\title{Wrapper around PCA}
\usage{
run_pca(so, approx_pca.flag = F, pca.assay = "RNA", n_pcs = 30, verbose = F)
}
\arguments{
\item{so}{a SeuratObject of your data}

\item{approx_pca.flag}{approximate PCA calculation?}

\item{pca.assay}{which assay to use, SCT or RNA}

\item{verbose}{whether to be verbose with output}
}
\description{
functionality for returning PCA results in Seurat Object
}
