% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{sample.per.label}
\alias{sample.per.label}
\title{Sample cells per label}
\usage{
sample.per.label(labels, size, boolean.flag = T, v, remove.flag = F)
}
\arguments{
\item{labels}{list of cells based on labels}

\item{size}{number of cells to sample per label}

\item{boolean.flag}{not sure}

\item{v}{not sure what this is}

\item{remove.flag}{not sure}
}
\value{
a list of cell names that correspond to the subsample
}
\description{
samples \emph{size} number of cells from each
labelled population
}
