% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{scale_and_center}
\alias{scale_and_center}
\title{Scale and Center}
\usage{
scale_and_center(X, MARGIN)
}
\arguments{
\item{X}{a numeric matrix-like object}

\item{MARGIN}{margin}
}
\value{
a scaled and centered matrix
}
\description{
scale and center a matrix
}
