% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_assignments.R
\name{scRNA_cluster.annotation.HG}
\alias{scRNA_cluster.annotation.HG}
\title{Hypergeometric Test}
\usage{
scRNA_cluster.annotation.HG(r, cell.clusters, EM.flag)

scRNA_cluster.annotation.HG(r, cell.clusters, EM.flag)
}
\arguments{
\item{r}{data list including the gene expression data}

\item{cell.clusters}{a \emph{n}-length list of cluster assignments, where \emph{n} = number of cells}

\item{EM.flag}{whether to compute EM}
}
\value{
data list, but now including a \code{cell.types} slot

data list, but now including a \code{cell.types} slot
}
\description{
function for computing hypergeometric test
on single cell clustered data

function for computing hypergeometric test
on single cell clustered data
}
