% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_assignments.R
\name{scRNA_cluster.annotation.ttest}
\alias{scRNA_cluster.annotation.ttest}
\title{Compute \emph{t}-tests}
\usage{
scRNA_cluster.annotation.ttest(r, cell.clusters, minZ = 10, subsample = F)
}
\arguments{
\item{r}{data list including the gene expression data}

\item{cell.clusters}{a \emph{n}-length list of cluster assignments, where \emph{n} = number of cells}

\item{minZ}{cutoff for significance in the -log10(p) space (default = 10)}
}
\value{
data list, but now including \code{cell.types}, and \code{assignments.ttests} slots
}
\description{
function for computing two sample \emph{t}-test
on single cell clustered data
}
