% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_assignments.R
\name{scRNA_markers_assign}
\alias{scRNA_markers_assign}
\title{scRNA Marker Based Cell Type Assignments}
\usage{
scRNA_markers_assign(
  r,
  cell.clusters,
  cell.sig,
  immune.markers,
  non.immune.cell.types = c("Endothelial", "CAF", "Malignant"),
  EM.flag = F,
  OE.type = "V1",
  test.type = "ttest",
  minZ = 10
)
}
\arguments{
\item{r}{data list including the gene expression data}

\item{cell.clusters}{a \emph{n}-length list to where \emph{n} is the number of cells}

\item{cell.sig}{cell type signatures (list of lists of markers associated with each cell-type, default exists)}

\item{immune.markers}{immune markers (default exists)}

\item{non.immune.cell.types}{which cell types are not immune cell types?}

\item{EM.flag}{run EM? (default = F)}

\item{OE.type}{which version of overall-expression to calculate (default = "V1")}

\item{test.type}{which kind of test to test marker significance between clusters}

\item{minZ}{minimum significance cutoff to assign a cell-type (default = 10)}
}
\value{
a list with objects named "c" and "m" corresponding to your subsampled counts and metadata
}
\description{
Using gene expression markers, unsupervised clusters,
and gene expression, assign the likely celltype to those
clusters
}
