% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{seuratify}
\alias{seuratify}
\title{This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.}
\usage{
seuratify(counts, prj_name, metadata)
}
\arguments{
\item{counts}{a \emph{m} x \emph{n} matrix of raw count data where \emph{m} = no. genes, and \emph{n} = no. cells}

\item{prj_name}{project name string}

\item{metadata}{matrix of metadata with at least a column cell identifier named "cellid"}
}
\value{
a Seurat object of your count matrix with metadata
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
