% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageProcessing.R
\name{spatial_sample_visualization}
\alias{spatial_sample_visualization}
\title{Spatial Cell Mask Visualization}
\usage{
spatial_sample_visualization(
  seg_path,
  celltypes,
  cell2rgb,
  samplename,
  background = "black",
  outpath = "~/",
  file_prefix = "",
  cont_field = "",
  contvals = NULL
)
}
\arguments{
\item{seg_path}{string of full path to cell segmentation map}

\item{celltypes}{a list of cell labels with names of the list corresponding to cellid}

\item{cell2rgb}{a dictionary that matches discrete cell label to a rgb code}

\item{samplename}{string name of SMI sample}

\item{background}{(default "black") background color ("black" or "white")}

\item{outpath}{(default = "~/") string path to output directory}

\item{file_prefix}{(default = "") string prefix of file}

\item{cont_field}{the celltype for which the coloring is continuous, not discrete. names must correspond to cellid}

\item{contvals}{list of the hexcolors corresponding to each cell to be plotted on continuous scale}
}
\description{
Plots for each sample, an image of each segmented
cell colored by user specified features
}
