% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{spearman.cor}
\alias{spearman.cor}
\title{Compute Spearman Correlation on two list objects}
\usage{
spearman.cor(
  v1,
  v2 = NULL,
  method = "spearman",
  use = "pairwise.complete.obs",
  match.flag = F,
  alternative = "two.sided",
  upper.tri.flag = F
)
}
\arguments{
\item{v1}{matrix, rows are samples, columns are genes}

\item{v2}{metric to evaluate the correlation with (must be same length as rows of v1)}

\item{method}{(default = "spearman")}

\item{use}{(default = "pairwise.complete.obs") type of correlation}

\item{match.flag}{(default = F)}

\item{alternative}{(default = "two.sided") alternative hypothesis for correlation test}

\item{upper.tri.flag}{(default = F)}
}
\value{
a new subsetted list object
}
\description{
Wrapper around correlation.
}
