% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{subset_list}
\alias{subset_list}
\title{Subset a list object based on cell ids}
\usage{
subset_list(r, subcells)
}
\arguments{
\item{r}{original list object}

\item{subcells}{ids of cells to subset}
}
\value{
a new subsetted list object
}
\description{
Use the cell ids to make a new list object
with cells subsets.
}
