% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{t.test.mat}
\alias{t.test.mat}
\title{Compute \emph{t}-tests}
\usage{
\method{t}{test.mat}(m, b, two.sided = F, rankf = F, fold.changeF = F)
}
\arguments{
\item{m}{markers matrix (subsampled, in some cases)}

\item{b}{boolean for the right cluster}

\item{two.sided}{boolean to determine whether you want to run a two-sided t-test}

\item{rankf}{(default = F)}

\item{fold.changeF}{(default = F)}
}
\value{
p-value table
}
\description{
computes \emph{t}-tests for a matrix
}
