% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{transfer_data_list_to_so}
\alias{transfer_data_list_to_so}
\title{Transfer meta data from a list object to a
Seurat object.}
\usage{
transfer_data_list_to_so(
  r,
  so,
  transfer_list = c("coor", "samples", "TMAs", "patients", "samples", "sites",
    "treatment", "cell.types", "cell.types2")
)
}
\arguments{
\item{r}{original list object}

\item{so}{seurat object}

\item{transfer_list}{list of meta fields to transfer}
}
\value{
a new subsetted list object
}
\description{
User can specify which fields to transfer from
a list object to a seurat object
}
