% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_wrappers.R
\name{umap_tsne}
\alias{umap_tsne}
\title{Wrapper around UMAP and/or TSNE}
\usage{
umap_tsne(
  so,
  tsne.flag = F,
  umap.flag = F,
  n_dims = 30,
  reduction = "pca",
  verbose = F
)
}
\arguments{
\item{so}{a SeuratObject of your data}

\item{tsne.flag}{boolean, run tSNE?}

\item{umap.flag}{boolean, run umap?}

\item{n_dims}{number of dimensions you want to use from your embedding space}

\item{reduction}{string that matches to the right reduction slot}

\item{verbose}{whether to be verbose in the calculation}
}
\value{
SeuratObject with computed visualizations
}
\description{
functionality for returning UMAP an TSNE results
}
