% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_wrappers.R
\name{var_explained_seurat}
\alias{var_explained_seurat}
\title{Variance Explained}
\usage{
var_explained_seurat(so, slot = "scale.data")
}
\arguments{
\item{so}{a SeuratObject of your data (with pca run)}

\item{slot}{which data slot to use}
}
\value{
a list of all parameters calculated
}
\description{
extract variance explained from Seurat Data
}
