% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{run_pca}
\alias{run_pca}
\title{Wrapper around PCA}
\usage{
run_pca(so, approx_pca.flag = F, pca.assay = "RNA", n_pcs = 30)
}
\arguments{
\item{so}{a SeuratObject of your data}

\item{approx_pca.flag}{approximate PCA calculation?}

\item{pca.assay}{= which assay to use, SCT or RNA}
}
\description{
functionality for returning PCA results in Seurat Object
}
