% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R, R/seurat_wrappers.R
\name{seuratify}
\alias{seuratify}
\title{Seuratify your Single Cell Data}
\usage{
seuratify(counts, prj_name = "")

seuratify(counts, prj_name = "")
}
\arguments{
\item{counts}{a \emph{m} x \emph{n} matrix of raw count data where \emph{m} = no. genes, and \emph{n} = no. cells}

\item{prj_name}{project name string}

\item{metadata}{matrix of metadata with at least a column cell identifier named "cellid"}
}
\value{
a Seurat object of your count matrix with metadata

a Seurat object of your count matrix with metadata
}
\description{
This function takes in raw counts data and
associated metadata and creates a SeuratObject

This function takes in raw counts data and
associated metadata and creates a SeuratObject
}
