% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R, R/seurat_wrappers.R
\name{std_preprocess}
\alias{std_preprocess}
\title{Standard single cell RNA preprocessing}
\usage{
std_preprocess(so, nfeatures = 100)

std_preprocess(so, nfeatures = 100)
}
\arguments{
\item{so}{a SeuratObject of your data}
}
\value{
a preprocessed SeuratObject

a preprocessed SeuratObject
}
\description{
normalize, transform, and scale

normalize, transform, and scale
}
